backend/requirements.txt
eth-brownie
Django ==4.1.12
django-filter
django-extensions
djangorestframework
django-cors-headers
psycopg2-binary
matplotlib
django-haystack
